+++ src/System.Buffers.Primitives/System.Buffers.Primitives.csproj
@@ -8,6 +8,7 @@
     <PackageIconUrl>http://go.microsoft.com/fwlink/?linkid=833199</PackageIconUrl>
   </PropertyGroup>
   <ItemGroup>
+    <PackageReference Include="System.Buffers" Version="4.3.0" />
     <PackageReference Include="System.Memory" Version="4.4.0-*" />
     <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.4.0-*" />
     <PackageReference Include="System.Diagnostics.Tools" Version="4.3.0" />
