+++ src/System.Buffers.Primitives/System/Buffers/NativeBufferPool.cs
